+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -5,8 +5,7 @@
 ### Features Added
 - Added support for service version 2021-06-08.
 - Added a new paginated method for listing page ranges, `list_page_ranges()`. This replaces `get_page_ranges()` which has been deprecated.
-- Added support for `copy_source_tags` to `start_copy_from_url()` which allows callers to specify whether blob tags
-  should be replaced or copied during a synchronous copy operation.
+- Added support for copying source blob tags with `start_copy_from_url()` by specifying `"COPY"` for the `tags` keyword. 
 - Added support for `create_container_if_not_exists()` for `BlobContainerClient`
 
 ## 12.11.0 (2022-03-29)